refactor: breakout config into its own file for readability
